+++ src/Compiler.cpp
@@ -5,7 +5,7 @@
  *      Author: skutnii
  */
 
-#include "Compiler.h"
+#include "Compiler.hpp"
 
 namespace dirac {
