+++ src/main.cpp
@@ -6,7 +6,8 @@
  */
 
 #include <iostream>
-#include "eval.h"
+
+#include "eval.hpp"
 
 using namespace dirac;
